+++ application/views/adminListBook.php
@@ -37,7 +37,7 @@
             <td><?php echo $book["Book"]["thumbnail_address"];?></td>
             <td><?php echo $book["Book"]["bookPDF"];?></td>
             <td><a href="<?php echo LINK; ?>/admin/editBook/<?php echo $book["Book"]["book_id"];?>"><button>Edit</button></a></td>
-            <td><a href="<?php echo LINK; ?>/admin/deleteBook/<?php echo $book["Book"]["book_id"];?>"><button>Delete</button></a></td>
+            <td><a onclick="return confirm('Are you sure you want to delete this book')" href="<?php echo LINK; ?>/admin/deleteBook/<?php echo $book["Book"]["book_id"];?>"><button>Delete</button></a></td>
         </tr>
         <?php
         }
